it seems to work but... no!

+++ problem128/Solution128.java
@@ -1,12 +1,13 @@
 public class Solution128 {
 
 	public static void main(String[] args) {
+
 		// For timing
 		long t1 = System.currentTimeMillis();
 
 		Node node = new Node();
 
-		int sol = node.current.id;
+		long sol = node.current.id;
 		System.out.println("The solution is: " + sol);
 		System.out.println("Took: " + (System.currentTimeMillis() - t1) + "ms");
 	}
